[refactor] dotenv 설치
