+++ .env.local
@@ -1,0 +1,2 @@
+ELASTIC_SEARCH_URL=http://localhost:9200
+MONGODB_URL=mongodb://rainbow:rainbowA307@127.0.0.1:27017
