+++ .gitignore
@@ -202,7 +202,7 @@ web_modules/
 .env.development.local
 .env.test.local
 .env.production.local
-.env.local
+!.env.local
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
