+++ package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "@elastic/elasticsearch": "^8.10.0",
     "crypto": "^1.0.1",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "ioredis": "^5.3.2",
     "mongodb": "^6.1.0",
