+++ public/common/elasticSearch/insertData.js
@@ -1,11 +1,12 @@
 import { Client } from "@elastic/elasticsearch";
+import { ELASTIC_SEARCH_URL } from "../tools/config";
 
-const esClient = new Client({ node: "http://localhost:9200" });
+const esClient = new Client({ node: ELASTIC_SEARCH_URL });
 
 const indexName = "themes";
 
 const insertData = async (data) => {
-  const body = data.flatMap((doc) => [{ index: { _index: "indexName" } }, doc]);
+  const body = data.flatMap((doc) => [{ index: { _index: indexName } }, doc]);
   await esClient.bulk({ refresh: true, body });
 };
