+++ public/common/tools/config.js
@@ -1,0 +1,2 @@
+export const ELASTIC_SEARCH_URL = process.env.ELASTIC_SEARCH_URL;
+export const MONGODB_URL = process.env.MONGODB_URL;
