+++ public/common/tools/fetch.js
